+++ lib/cache.js
@@ -455,12 +455,19 @@ function addRemoteGit (u, parsed, name, silent, cb_) {
   })
 }
 
+var cloneGitRemote = require("./git-cache.js").cloneGitRemote
 function checkGitDir (p, u, co, origUrl, silent, cb) {
+  // xxx git-module refactor
+  var g = npm.config.get("git")
+    , t = npm.tmp
+    , c = npm.cache
+
   fs.stat(p, function (er, s) {
-    if (er) return cloneGitRemote(p, u, co, origUrl, silent, cb)
+    if (er) return cloneGitRemote(p, u, co, origUrl, silent
+      , c, t, g, log, gitDoneCb, cb)
     if (!s.isDirectory()) return rm(p, function (er){
       if (er) return cb(er)
-      cloneGitRemote(p, u, co, origUrl, silent, cb)
+      cloneGitRemote(p, u, co, origUrl, silent, c, t, g, log, gitDoneCb, cb)
     })
 
     var git = npm.config.get("git")
@@ -478,9 +485,10 @@ function checkGitDir (p, u, co, origUrl, silent, cb) {
         if (er || u !== stdout.trim()) {
           log.warn( "`git config --get remote.origin.url` returned "
                   + "wrong result ("+u+")", stdoutTrimmed )
-          return rm(p, function (er){
+          return rm(p, function (er) {
             if (er) return cb(er)
-            cloneGitRemote(p, u, co, origUrl, silent, cb)
+            cloneGitRemote(p, u, co, origUrl, silent
+              , c, t, g, log, gitDoneCb, cb)
           })
         }
         log.verbose("git remote.origin.url", stdoutTrimmed)
@@ -490,38 +498,8 @@ function checkGitDir (p, u, co, origUrl, silent, cb) {
   })
 }
 
-function cloneGitRemote (p, u, co, origUrl, silent, cb) {
-  mkdir(p, function (er) {
-    if (er) return cb(er)
-
-    var git = npm.config.get("git")
-    var args = [ "clone", "--mirror", u, p ]
-    var env = gitEnv()
-
-    // check for git
-    which(git, function (err) {
-      if (err) {
-        err.code = "ENOGIT"
-        return cb(err)
-      }
-      exec(git, args, {cwd: p, env: env}, function (er, stdout, stderr) {
-        stdout = (stdout + "\n" + stderr).trim()
-        if (er) {
-          if (silent) {
-            log.verbose("git clone " + u, stdout)
-          } else {
-            log.error("git clone " + u, stdout)
-          }
-          return cb(er)
-        }
-        log.verbose("git clone " + u, stdout)
-        archiveGitRemote(p, u, co, origUrl, cb)
-      })
-    })
-  })
-}
-
 function gitDoneCb (tmp, resolved, cb) {
+  // xxx git-module refactor
   addLocalTarball(tmp, function (er, data) {
     if (data) data._resolved = resolved
     cb(er, data)
@@ -529,12 +507,13 @@ function gitDoneCb (tmp, resolved, cb) {
 }
 
 function archiveGitRemote (p, u, co, origUrl, cb) {
-  var git = npm.config.get("git")
+  // xxx git-module refactor
+  var g = npm.config.get("git")
     , t = npm.tmp
     , c = npm.cache
 
   require("./git-cache.js")
-    .archiveGitRemote(p, u, co, origUrl, c, t, git, log, gitDoneCb, cb)
+    .archiveGitRemote(p, u, co, origUrl, c, t, g, log, gitDoneCb, cb)
 }
 
 var gitEnv_
