+++ lib/git-cache.js
@@ -8,10 +8,41 @@ var chownr = require("chownr")
   , once = require("once")
   , fs = require("graceful-fs")
   , spawn = require("child_process").spawn
+  , which = require("which")
 
+exports.cloneGitRemote = cloneGitRemote
+function cloneGitRemote (p, u, co, origUrl, silent, ca, t, git, log, gitDoneCb, cb) {
+  mkdir(p, function (er) {
+    if (er) return cb(er)
+
+    var args = [ "clone", "--mirror", u, p ]
+    var env = gitEnv()
+
+    // check for git
+    which(git, function (err) {
+      if (err) {
+        err.code = "ENOGIT"
+        return cb(err)
+      }
+      exec(git, args, {cwd: p, env: env}, function (er, stdout, stderr) {
+        stdout = (stdout + "\n" + stderr).trim()
+        if (er) {
+          if (silent) {
+            log.verbose("git clone " + u, stdout)
+          } else {
+            log.error("git clone " + u, stdout)
+          }
+          return cb(er)
+        }
+        log.verbose("git clone " + u, stdout)
+        archiveGitRemote(p, u, co, origUrl, ca, t, git, log, gitDoneCb, cb)
+      })
+    })
+  })
+}
 
 exports.archiveGitRemote = archiveGitRemote
-function archiveGitRemote (p, u, co, origUrl, cache, tmp, git, log, gitDoneCb, cb) {
+function archiveGitRemote (p, u, co, origUrl, ca, t, git, log, gitDoneCb, cb) {
   var archive = [ "fetch", "-a", "origin" ]
   var resolve = [ "rev-list", "-n1", co ]
   var env = gitEnv()
@@ -28,7 +59,7 @@ function archiveGitRemote (p, u, co, origUrl, cache, tmp, git, log, gitDoneCb, c
       return cb(er)
     }
     log.verbose("git fetch -a origin ("+u+")", stdout)
-    tmp = path.join(tmp, Date.now()+"-"+Math.random(), "tmp.tgz")
+    tmp = path.join(t, Date.now()+"-"+Math.random(), "tmp.tgz")
     verifyOwnership()
   })
 
@@ -37,7 +68,7 @@ function archiveGitRemote (p, u, co, origUrl, cache, tmp, git, log, gitDoneCb, c
       log.silly("verifyOwnership", "skipping for windows")
       resolveHead()
     } else {
-      getCacheStat(cache, log, function(er, cs) {
+      getCacheStat(ca, log, function(er, cs) {
         if (er) {
           log.error("Could not get cache stat")
           return cb(er)
